Split up http_handlers.rs
