+++ aggregator/src/aggregator/http_handlers/tests/mod.rs
@@ -1,0 +1,6 @@
+mod aggregate_share;
+mod aggregation_job_continue;
+mod aggregation_job_init;
+mod collection_job;
+mod hpke_config;
+mod report;
